requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0
python-dotenv==1.0.1
googlenewsdecoder>=0.1.2
pyyaml>=6.0.1
